Fixed it to delete data before calling XDestroyImage (necessary for WIN32 to insure the same routines create and free the data memory). Added Xt_Warning_handler and Xt_Error_Handler.  Changed library order to prevent zero height/width problem with 3.14 on Solaris.  Fixed it to not realize the top-level shell or destroy it on close.  This should not be necessary, but neither is needed, and it resolves strange problems on several platforms.  Fixed it to do Strip_clear in dlgrequest_quit, so ca_clear_channel would be called and prevent disconnect messages on close in 3.14 (as well as being good practice).

+++ StripDataSource.c
@@ -455,7 +455,7 @@ StripDataSource_min_max (StripDataSourceInfo *sds, struct timeval h0,
 	    }
 #else
 	    /* First part is first to end */
-	    for(i=first; i < sds->buf_size; i++)
+	    for(i=first; i < (int)sds->buf_size; i++)
 	    {
 		if(cd->val[i] < min) min=cd->val[i]; 
 		if(cd->val[i] > max) max=cd->val[i];
